refactor identity center into lib
